About section, helper text, design improvements
The about section is reachable over the entry in the menu.
It is implemented as a dialog and shows a text giving information
about the project and liks to relevant pages. A know issue is
that the links are not clickable (yet).

There is a helper text below the input text field now, giving
information about the current operation, encoding or decoding.
The app now uses edge to edge layout, which until now gets only
rid of the brighly colored task bar, but might be improved in the
future to show scrolled elements behind the task bar.